Fixed #184. This issue was caused when a given definition is used in multiple places within a model. The a copy of the constraint will be added for each time the constraint target matches one of those locations.
