+++ core/src/main/java/gov/nist/secauto/metaschema/core/model/validation/JsonSchemaContentValidator.java
@@ -34,7 +34,8 @@ public class JsonSchemaContentValidator
   private final Schema schema;
 
   /**
-   * Construct a new JSON schema validator using the provided reader to load the JSON schema.
+   * Construct a new JSON schema validator using the provided reader to load the
+   * JSON schema.
    *
    * @param reader
    *          the JSON schema reader
@@ -44,7 +45,8 @@ public class JsonSchemaContentValidator
   }
 
   /**
-   * Construct a new JSON schema validator using the provided input stream to load the JSON schema.
+   * Construct a new JSON schema validator using the provided input stream to load
+   * the JSON schema.
    *
    * @param is
    *          the JSON schema input source
@@ -54,7 +56,8 @@ public class JsonSchemaContentValidator
   }
 
   /**
-   * Construct a new JSON schema validator using the provided JSON object for the JSON schema.
+   * Construct a new JSON schema validator using the provided JSON object for the
+   * JSON schema.
    *
    * @param jsonSchema
    *          the JSON schema
@@ -64,7 +67,8 @@ public class JsonSchemaContentValidator
   }
 
   /**
-   * Construct a new JSON schema validator using the provided JSON tokenizer to load the schema.
+   * Construct a new JSON schema validator using the provided JSON tokenizer to
+   * load the schema.
    *
    * @param tokenizer
    *          the JSON schema token stream
@@ -121,8 +125,8 @@ public class JsonSchemaContentValidator
    * Build validation findings from a validation exception.
    *
    * @param exception
-   *          the JSON schema validation exception generated during schema validation representing the
-   *          issue
+   *          the JSON schema validation exception generated during schema
+   *          validation representing the issue
    * @param resourceUri
    *          the resource the issue was found in
    * @return the stream of findings
@@ -139,7 +143,8 @@ public class JsonSchemaContentValidator
   }
 
   /**
-   * Records an identified individual validation result found during JSON schema validation.
+   * Records an identified individual validation result found during JSON schema
+   * validation.
    */
   public static class JsonValidationFinding implements IValidationFinding {
     @NonNull
@@ -148,12 +153,12 @@ public class JsonSchemaContentValidator
     private final URI documentUri;
 
     /**
-     * Construct a new XML schema validation finding, which represents an issue identified during XML
-     * schema validation.
+     * Construct a new XML schema validation finding, which represents an issue
+     * identified during XML schema validation.
      *
      * @param exception
-     *          the JSON schema validation exception generated during schema validation representing the
-     *          issue
+     *          the JSON schema validation exception generated during schema
+     *          validation representing the issue
      * @param resourceUri
      *          the resource the issue was found in
      */
