+++ core/src/test/resources/content/issue184-constraints.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?xml-model href="../../../../metaschema/schema/xml/metaschema-meta-constraints.xsd" type="application/xml" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<metaschema-meta-constraints xmlns="http://csrc.nist.gov/ns/oscal/metaschema/1.0">
+    <context>
+        <metapath target="//*"/>
+        <constraints>
+			<allowed-values target="@value">
+                <enum value="value1">Value #1</enum>
+			</allowed-values>
+        </constraints>
+    </context>
+</metaschema-meta-constraints>
